refactor and css fixes
